webpage/requirements.txt
loguru==0.7.0
numpy==1.24.2
Pillow==9.5.0
plotly==5.8.0
python-keycloak==2.16.1
scipy==1.10.1
sentry-sdk==1.20.0
streamlit==1.21.0
streamlit-keycloak==1.1.1
urllib3==1.26.15
